Edit the file in DB
