+++ src/model/queries/business.js
@@ -1,6 +1,5 @@
 const dbConnection = require('../database/db_connection');
 
-
 const getSearch = name => new Promise((resolve, reject) => {
   const sql = {
     text: 'SELECT * FROM business WHERE name LIKE $1',
@@ -8,7 +7,7 @@ const getSearch = name => new Promise((resolve, reject) => {
   };
   dbConnection.query(sql, (error, res) => {
     if (error) return reject(error);
-    resolve(res.rows);
+    return resolve(res.rows);
   });
 });
 
@@ -19,29 +18,6 @@ const getSearchResult = name => new Promise((resolve, reject) => {
   };
   dbConnection.query(sql, (error, res) => {
     if (error) return reject(error);
-    resolve(res.rows);
-  });
-});
-
-const getReviews = () => new Promise((resolve, reject) => {
-  const sql = {
-    text: 'SELECT * FROM review ORDER BY id desc',
-  };
-  dbConnection.query(sql, (error, res) => {
-    if (error) return reject(error);
-    resolve(res.rows);
-  });
-});
-
-const checkUser = email => new Promise((resolve, reject) => {
-  const sql = {
-    text: 'SELECT * FROM users WHERE email=$1;',
-    values: [email],
-  };
-  dbConnection.query(sql, (error, res) => {
-    if (error) {
-      return reject(new Error('Error in DB'));
-    }
     return resolve(res.rows);
   });
 });
@@ -49,6 +25,4 @@ const checkUser = email => new Promise((resolve, reject) => {
 module.exports = {
   getSearch,
   getSearchResult,
-  getReviews,
-  checkUser,
 };
